requirements.txt
Pillow>=10.0.0
imageio-ffmpeg>=0.5.1

